requirements.txt
Flask==2.2.2
flask-cors==3.1.1
numpy==1.24.2
pandas==1.5.3
tensorflow==2.11.0
transformers==4.32.0
Pillow==9.3.0
torch==2.0.1
openai==0.27.1
